fix: move and modify read_csv function
