+++ preprocessing/workflows/hydranl_waterlevel_workflow.py
@@ -2,7 +2,7 @@ import os
 from pathlib import Path
 import pandas as pd
 from preprocessing.step2_mechanism_data.hydranl_read import HydraNLReadWaterLevel
-from preprocessing.workflows.write_database_workflow import read_csv_linesep
+from preprocessing.common_functions import read_csv_linesep
 
 import logging
